requirements.txt
click==8.0.3
joblib==1.1.0
loguru==0.6.0
paho-mqtt==1.6.1
pandas==1.4.0
psutil==5.9.0
psycopg2-binary==2.9.3
pyyaml==6.0
requests==2.27.1
scikit-learn==1.0.2
